Many namespace fixes on generators.
