+++ lib/rails/generators/neo4j/upgrade_v8/templates/migration.erb
@@ -1,8 +1,8 @@
 class <%= @migration_class_name %> < Neo4j::Migrations::Base
   def up
     <% @schema.each do |type, data| %>
-      <% data.each do |element| %>
-    add_<%= type %> <%= element[:model].inspect %>, <%= element[:property].inspect %>, force: true
+      data.each do |element| %>
+    add_<%= type %> :<%= element[:label].inspect %>, <%= element[:property_name].inspect %>, force: true
       <% end %>
     <% end %>
   end
@@ -10,7 +10,7 @@ class <%= @migration_class_name %> < Neo4j::Migrations::Base
   def down
     <% @schema.each do |type, data| %>
       <% data.each do |element| %>
-    drop_<%= type %> <%= element[:model].inspect %>, <%= element[:property].inspect %>
+    drop_<%= type %> <%= element[:label].inspect %>, <%= element[:property_name].inspect %>
       <% end %>
     <% end %>
   end
